All endpoints protect appropriately.

+++ controllers/Items.py
@@ -1,8 +1,14 @@
 from flask import jsonify, request
 from controllers import *
 from models import Items
+import jwt
+from flask_jwt_extended import (
+    JWTManager, jwt_required, create_access_token,
+    get_jwt_identity
+)
 
 @controllers.route('/items', methods=['GET'])
+@jwt_required(optional=False)
 def get_items():
 
     items = Items.query.all()
@@ -25,6 +31,7 @@ def get_items():
 
 
 @controllers.route('/items/<int:id>', methods=['GET'])
+@jwt_required(optional=False)
 def get_item(id):
     try:
         item = Items.query.get(id)
@@ -53,38 +60,49 @@ def get_item(id):
         return jsonify(data)
 
 @controllers.route('/items/<int:id>', methods=['PUT'])
+@jwt_required(optional=False)
 def update_item(id):
-    try:
 
-        args = request.get_json()
+    data = get_jwt_identity()
 
-        location = args['location']
-        ownerId = args['ownerId']
-        name = args['name']
-        description = args['description']
-        imageURL = args['imageURL']
-        rating = args['rating']
+    if Permissions.get_permission(data['permission']).permission == "Admin" or Users.query.filter_by(email=data['email']).first().id == Items.get_item(id).ownerId:
+        try:
 
-        if not Items.get_item(id):
+            args = request.get_json()
+
+            location = args['location']
+            ownerId = args['ownerId']
+            name = args['name']
+            description = args['description']
+            imageURL = args['imageURL']
+            rating = args['rating']
+
+            if not Items.get_item(id):
+                return jsonify(isError=True,
+                        message="Could not find item",
+                        statusCode=404,
+                        data=str("Not Found")), 404
+            else:
+                Items.update_item(id, location, ownerId, name, description, imageURL, rating)
+
+        except Exception as e:
             return jsonify(isError=True,
-                       message="Could not find item",
-                       statusCode=404,
-                       data=str("Not Found")), 404
+                        message="Error",
+                        statusCode=500,
+                        data=str("Internal Server Error")), 500
         else:
-            Items.update_item(id, location, ownerId, name, description, imageURL, rating)
-
-    except Exception as e:
-        return jsonify(isError=True,
-                       message="Error",
-                       statusCode=500,
-                       data=str("Internal Server Error")), 500
+            return jsonify(isError=False,
+                        message="Success",
+                        statusCode=201,
+                        data=name), 201
     else:
-        return jsonify(isError=False,
-                       message="Success",
-                       statusCode=201,
-                       data=name), 201
+        return jsonify(isError=True,
+                    message="You are Unauthorized",
+                    statusCode=401,
+                    data=str("Restricted access")), 401
 
 @controllers.route('/items', methods=['POST'])
+@jwt_required(optional=False)
 def post_item():
     try:
         args = request.get_json()
@@ -110,24 +128,34 @@ def post_item():
                        data=name), 201
 
 @controllers.route('/items/<int:id>', methods=['DELETE'])
+@jwt_required(optional=False)
 def delete_item(id):
-    try:
 
-        if not Items.get_item(id):
+    data = get_jwt_identity()
+
+    if Permissions.get_permission(data['permission']).permission == "Admin" or Users.query.filter_by(email=data['email']).first().id == Items.get_item(id).ownerId:
+        try:
+
+            if not Items.get_item(id):
+                return jsonify(isError=True,
+                        message="Could not find item",
+                        statusCode=404,
+                        data=str("Not Found")), 404
+            else:
+                Items.delete_item(id)
+
+        except Exception as e:
             return jsonify(isError=True,
-                       message="Could not find item",
-                       statusCode=404,
-                       data=str("Not Found")), 404
+                        message="Error",
+                        statusCode=500,
+                        data=str("Item deletion error")), 500
         else:
-            Items.delete_item(id)
-
-    except Exception as e:
-        return jsonify(isError=True,
-                       message="Error",
-                       statusCode=500,
-                       data=str("Item deletion error")), 500
+            return jsonify(isError=False,
+                        message="Success",
+                        statusCode=201,
+                        data="item"), 201
     else:
-        return jsonify(isError=False,
-                       message="Success",
-                       statusCode=201,
-                       data="item"), 201
+        return jsonify(isError=True,
+                    message="You are Unauthorized",
+                    statusCode=401,
+                    data=str("Restricted access")), 401
